// tina/__generated__/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** References another document, used as a foreign key */
  Reference: { input: any; output: any; }
  JSON: { input: any; output: any; }
  /** A rich-text document, renderable by TinaMarkdown */
  RichText: { input: TinaMarkdownContent; output: TinaMarkdownContent; }
};

export type SystemInfo = {
  __typename?: 'SystemInfo';
  filename: Scalars['String']['output'];
  title?: Maybe<Scalars['String']['output']>;
  basename: Scalars['String']['output'];
  hasReferences?: Maybe<Scalars['Boolean']['output']>;
  breadcrumbs: Array<Scalars['String']['output']>;
  path: Scalars['String']['output'];
  relativePath: Scalars['String']['output'];
  extension: Scalars['String']['output'];
  template: Scalars['String']['output'];
  collection: Collection;
};
// ...
export type Document = {
  id: Scalars['ID']['output'];
  _sys?: Maybe<SystemInfo>;
  _values: Scalars['JSON']['output'];
};
// ...
export type FlashDesigns = Node & Document & {
  __typename?: 'FlashDesigns';
  title: Scalars['String']['output'];
  price?: Maybe<Scalars['String']['output']>;
  category?: Maybe<Scalars['String']['output']>;
  image?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['RichText']['output']>;
  id: Scalars['ID']['output'];
  _sys: SystemInfo;
  _values: Scalars['JSON']['output'];
};

export type StringFilter = {
  startsWith?: InputMaybe<Scalars['String']['input']>;
  eq?: InputMaybe<Scalars['String']['input']>;
  exists?: InputMaybe<Scalars['Boolean']['input']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type ImageFilter = {
  startsWith?: InputMaybe<Scalars['String']['input']>;
  eq?: InputMaybe<Scalars['String']['input']>;
  exists?: InputMaybe<Scalars['Boolean']['input']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type RichTextFilter = {
  startsWith?: InputMaybe<Scalars['String']['input']>;
  eq?: InputMaybe<Scalars['String']['input']>;
  exists?: InputMaybe<Scalars['Boolean']['input']>;
};

export type FlashDesignsFilter = {
  title?: InputMaybe<StringFilter>;
  price?: InputMaybe<StringFilter>;
  category?: InputMaybe<StringFilter>;
  image?: InputMaybe<ImageFilter>;
  description?: InputMaybe<RichTextFilter>;
};
// ...
export const FlashDesignsPartsFragmentDoc = gql`
    fragment FlashDesignsParts on FlashDesigns {
  __typename
  title
  price
  category
  image
  description
}
    `;
// ...
export const FlashDesignsConnectionDocument = gql`
    query flashDesignsConnection($before: String, $after: String, $first: Float, $last: Float, $sort: String, $filter: FlashDesignsFilter) {
  flashDesignsConnection(
    before: $before
    after: $after
    first: $first
    last: $last
    sort: $sort
    filter: $filter
  ) {
    pageInfo {
      hasPreviousPage
      hasNextPage
      startCursor
      endCursor
    }
    totalCount
    edges {
      cursor
      node {
        ... on Document {
          _sys {
            filename
            basename
            hasReferences
            breadcrumbs
            path
            relativePath
            extension
          }
          id
        }
        ...FlashDesignsParts
      }
    }
  }
}
    ${FlashDesignsPartsFragmentDoc}`;